Server: update mypage etc.
